Moved MadFixed32 definition from simplemad_sys to simplemad
