+++ simplemad/Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "simplemad"
-version = "0.5.3"
+version = "0.6.0"
 authors = ["Benjamin Dykstra <bendykst@gmail.com>"]
 description = "An interface for libmad, the MPEG audio decoding library"
 repository = "https://github.com/bendykst/simple-mad.rs"
